+++ docs/changelog.md
@@ -2,6 +2,10 @@
 
 ## 0.4.3
 
+### Added
+
+- [#169](https://github.com/JoshKarpel/spiel/pull/169) The Textual application title and subtitle are now set dynamically from the Spiel deck name and slide title, respectively.
+
 ### Fixed
 
 - [#168](https://github.com/JoshKarpel/spiel/pull/168) The correct type for the `suspend` optional argument to slide-level keybinding functions is now available as `spiel.SuspendType`.